% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{rDivide}
\alias{rDivide}
\title{Dzielenie macierzy}
\usage{
rDivide(quasarcl, inputMatrix, divisorMatrix)
}
\arguments{
\item{quasarcl}{(externalptr) Wskażnik na środowisko OpenCL}

\item{inputMatrix}{(matrix) Macierz wejściowa}

\item{divisorMatrix}{(matrix) Macierz, przez której elementy dzielimy}
}
\value{
(matrix) wynik
}
\description{
Dzieli każdy element macierzy inputMatrix przez odpowiadający mu element 
macierzy divisorMatrix
}
