% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{rLog10}
\alias{rLog10}
\title{Logarytm dziesiętny}
\usage{
rLog10(quasarcl, inputMatrix)
}
\arguments{
\item{quasarcl}{(externalptr) Wskażnik na środowisko OpenCL}

\item{inputMatrix}{(matrix) Macierz wejściowa}
}
\value{
(matrix) wynik
}
\description{
Oblicza logarytm dzięsiętny dla każdego elementu macierzy wejściowej
}
