% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{rMinusMatrix}
\alias{rMinusMatrix}
\title{Odejmowanie macierzy}
\usage{
rMinusMatrix(quasarcl, inputMatrix, subtrahendMatrix)
}
\arguments{
\item{quasarcl}{(externalptr) Wskażnik na środowisko OpenCL}

\item{inputMatrix}{(matrix) Macierz wejściowa}

\item{subtrahendMatrix}{(matrix) Macierz odejmowana}
}
\value{
(matrix) wynik
}
\description{
Odejmuje od każdego elementu macierzy inputMatrix odpowiadający mu element
macierzy subtrahendMatrix
}
