% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{rMultiplyCol}
\alias{rMultiplyCol}
\title{Mnożenie kolumn macierzy przez wektor}
\usage{
rMultiplyCol(quasarcl, inputMatrix, vector)
}
\arguments{
\item{quasarcl}{(externalptr) Wskażnik na środowisko OpenCL}

\item{inputMatrix}{(matrix) Macierz wejściowa}

\item{vector}{(numeric) Wektor}
}
\value{
(matrix) wynik
}
\description{
Mnoży kolejne kolumny macierzy wejściowej przez kolejne wartości wektora
}
