% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{rTranspose}
\alias{rTranspose}
\title{Transpozycja macierzy}
\usage{
rTranspose(quasarcl, inputMatrix)
}
\arguments{
\item{quasarcl}{(externalptr) Wskażnik na środowisko OpenCL}

\item{inputMatrix}{(matrix) Macierz wejściowa}
}
\value{
(matrix) wynik
}
\description{
Transponuje macierz wejściową
}
